requirements.txt
appdirs==1.4.4
appnope==0.1.3
astor==0.8.1
asttokens==2.2.1
attrs==23.1.0
backcall==0.2.0
beautifulsoup4==4.12.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
comm==0.1.4
contourpy==1.1.0
cycler==0.11.0
Cython==3.0.0
debugpy==1.6.7
decorator==5.1.1
executing==1.2.0
fastjsonschema==2.18.0
fonttools==4.42.0
formulaic==0.6.4
frozendict==2.3.8
html5lib==1.1
idna==3.4
interface-meta==1.3.0
ipykernel==6.25.0
ipython==8.14.0
jedi==0.19.0
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
jupyter_client==8.3.0
jupyter_core==5.3.1
kaleido==0.2.1
kiwisolver==1.4.4
linearmodels==5.1
lxml==4.9.3
MarkupSafe==2.1.3
matplotlib==3.7.2
matplotlib-inline==0.1.6
multitasking==0.0.11
mypy-extensions==1.0.0
nbformat==5.9.2
nest-asyncio==1.5.7
nltk==3.8.1
numpy==1.25.2
packaging==23.1
pandas==2.0.3
parso==0.8.3
patsy==0.5.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.0
platformdirs==3.10.0
plotly==5.15.0
polars==0.18.12
prompt-toolkit==3.0.39
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==12.0.1
Pygments==2.15.1
pyhdfe==0.1.2
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
pyzmq==25.1.0
referencing==0.30.2
regex==2023.8.8
requests==2.31.0
rpds-py==0.9.2
scipy==1.11.1
seaborn==0.12.2
setuptools-scm==7.1.0
six==1.16.0
soupsieve==2.4.1
stack-data==0.6.2
statsmodels==0.14.0
tenacity==8.2.2
tornado==6.3.2
tqdm==4.66.0
traitlets==5.9.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
wcwidth==0.2.6
webencodings==0.5.1
wordcloud==1.9.2
wrapt==1.15.0
yfinance==0.2.27
